symctx: change size_of_code and size_of_data type from int to size_t
Conflicts:

	fa/symctx.cc
	fa/symctx.hh

+++ fa/symctx.cc
@@ -19,5 +19,5 @@
 
 #include "symctx.hh"
 
-int SymCtx::size_of_data_ptr = 0;
-int SymCtx::size_of_code_ptr = 0;
+size_t SymCtx::size_of_data_ptr = 0;
+size_t SymCtx::size_of_code_ptr = 0;
